\name{yang-package}
\alias{yang-package}
\alias{yang}
\docType{package}
\title{
Rectal cancer cells and radiotherapy response
}
\description{
GDS1887 data from GEOquery
35-training and 11-test samples were analyzed
}
\details{
\tabular{ll}{
Package: \tab yang\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-02-23\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Dongsheng 
Maintainer: Dongsheng <yangd@ccf.org>
}
\references{
http://www.ncbi.nlm.nih.gov/sites/GDSbrowser?acc=GDS1887
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
