\name{yang}
\alias{yang}
\docType{data}
\title{
Rectal cancer cells and radiotherapy response
}
\description{
GDS1887 DATA from GEOquery
Analysis of rectal cancer cells from patients prior to preoperative radiotherapy. 
Response to radiotherapy determined by histopathologic examination of resected specimens. 
Results identify gene markers for the characterization and prediction of the response to
radiotherapy in rectal cancer
}
\usage{data(yang)}
\format{ this dataframe included 46 variables and 12651 observations
  The format is:{
  	\item \code{SM79076}:   {counts (value type)}  
	\item \code{SM79077}:   {counts (value type)} 
	\item \code{SM79078}:   {counts (value type)} 
	\item \code{SM79079}:   {counts (value type)} 
	\item \code{SM79080}:   {counts (value type)}  
	\item \code{SM79081}:   {counts (value type)} 
	\item \code{SM79082}:   {counts (value type)} 
	\item \code{SM79083}:   {counts (value type)} 
	\item \code{SM79084}:   {counts (value type)} 
	\item \code{SM79085}:   {counts (value type)} 
	\item \code{SM79088}:   {counts (value type)} 
	\item \code{SM79089}:   {counts (value type)} 
	\item \code{SM79090}:   {counts (value type)} 
	\item \code{SM79091}:   {counts (value type)} 
	\item \code{SM79092}:   {counts (value type)} 
	\item \code{SM79093}:   {counts (value type)} 
	\item \code{SM79094}:   {counts (value type)} 
	\item \code{SM79095}:   {counts (value type)} 
	\item \code{SM79096}:   {counts (value type)} 
	\item \code{SM79097}:   {counts (value type)} 
	\item \code{SM79098}:   {counts (value type)} 
	\item \code{SM79099}:   {counts (value type)} 
	\item \code{SM79104}:   {counts (value type)} 
	\item \code{SM79105}:   {counts (value type)} 
	\item \code{SM79106}:   {counts (value type)} 
	\item \code{SM79107}:   {counts (value type)} 
	\item \code{SM79108}:   {counts (value type)} 
	\item \code{SM79109}:   {counts (value type)} 
	\item \code{SM79068}:   {counts (value type)}    
	\item \code{SM79069}:   {counts (value type)}    
	\item \code{SM79070}:   {counts (value type)}    
	\item \code{SM79071}:   {counts (value type)}    
	\item \code{SM79072}:   {counts (value type)}     
	\item \code{SM79075}:   {counts (value type)}    
	\item \code{SM79102}:   {counts (value type)}    
	\item \code{SM79086}:   {counts (value type)}    
	\item \code{SM79087}:   {counts (value type)}    
	\item \code{SM79100}:   {counts (value type)}     
	\item \code{SM79101}:   {counts (value type)}     
	\item \code{SM79110}:   {counts (value type)}    
	\item \code{SM79111}:   {counts (value type)}    
	\item \code{SM79112}:   {counts (value type)}     
	\item \code{SM79073}:   {counts (value type)}        
	\item \code{SM79074}:   {counts (value type)}         
	\item \code{SM79103}:   {counts (value type)}        
	\item \code{SM79113}:   {counts (value type)}                               
	}
}
\details{
Affymetrix Human Genome U95A Array.35-training and 11-test samples were analyzed
Organism: Homo sapiens
Experiment type: Expression profiling by array
}
\source{
http://www.ncbi.nlm.nih.gov/geo/gds/gds_browse.cgi?gds=1887
}
\references{
http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE3493
"Nucleic Acids Res. 2005 Jan 1;33 Database Issue:D562-6"
}
\examples{
data(yang)
str(yang) 
}
\keyword{datasets}

                                     